% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_daily.R
\name{data_daily}
\alias{data_daily}
\title{Daily aggregated data}
\usage{
data_daily(deviceId, begin, end, as_list = FALSE)
}
\arguments{
\item{deviceId}{integer. A device identification code (id).}

\item{begin}{character. Start date. Use the format \code{day/month/year}. Example: \code{20/09/2024}}

\item{end}{character. End date. Use the format \code{day/month/year}. Example: \code{20/09/2024}}

\item{as_list}{logical. If \code{TRUE} the function will return a list, otherwise, a tibble.}
}
\value{
a list or a tibble, depending on the \code{as_list} argument.
}
\description{
Daily aggregated data
}
