% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hourly.R
\name{data_hourly}
\alias{data_hourly}
\title{Hourly aggregated data}
\usage{
data_hourly(deviceId, begin, end, as_list = FALSE)
}
\arguments{
\item{deviceId}{integer. A device identification code (id).}

\item{begin}{character. Start date and time. Use the format \verb{day/month/year hour}. Example: \verb{20/09/2024 09}}

\item{end}{character. End date and time. Use the format \verb{day/month/year hour}. Example: \verb{20/09/2024 09}}

\item{as_list}{logical. If \code{TRUE} the function will return a list, otherwise, a tibble.}
}
\value{
a list or a tibble, depending on the \code{as_list} argument.
}
\description{
Hourly aggregated data
}
