% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sensor.R
\name{data_sensor}
\alias{data_sensor}
\title{Data from sensor}
\usage{
data_sensor(deviceId, sensor, time = NULL, timeMax = NULL)
}
\arguments{
\item{deviceId}{integer. A device identification code (id).}

\item{sensor}{integer. A sensor code, as informed on \code{\link[=device_sensors]{device_sensors()}}.}

\item{time}{character. Start date and time. If not provided, the function will return data from the last day. Use the format \verb{day/month/year hour}. Example: \verb{20/09/2024 09}}

\item{timeMax}{character. End date and time. If not provided, the function will return data from the last day. Use the format \verb{day/month/year hour}. Example: \verb{20/09/2024 09}}
}
\value{
a tibble.
}
\description{
Retrieve data from a sensor on a device
}
