% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_info.R
\name{device_info}
\alias{device_info}
\title{Device info}
\usage{
device_info(deviceId)
}
\arguments{
\item{deviceId}{integer. A device identification code (id).}
}
\value{
a list.
}
\description{
Retrieve information for a specific device, including last updated data for a dashboard.
}
