% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_last_update.R
\name{device_last_update}
\alias{device_last_update}
\title{Device last update time}
\usage{
device_last_update(deviceId)
}
\arguments{
\item{deviceId}{integer. A device identification code (id).}
}
\value{
a date.
}
\description{
Retrieve device last update time
}
