% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_list.R
\name{device_list}
\alias{device_list}
\title{Device list}
\usage{
device_list()
}
\value{
a list.
}
\description{
Returns a list of available devices, including device's information and last updated data for a dashboard.
}
