% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_sensors.R
\name{device_sensors}
\alias{device_sensors}
\title{Device sensors}
\usage{
device_sensors(deviceId)
}
\arguments{
\item{deviceId}{integer. A device identification code (id).}
}
\value{
a tibble.
}
\description{
Retrieve sensors available from a device id.
}
