% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login on Plugfield}
\usage{
login(username = NULL, password = NULL, x_api_key = NULL)
}
\arguments{
\item{username}{character. Username, usually an e-mail.}

\item{password}{character.}

\item{x_api_key}{character. Provided by Plugfield.}
}
\value{
no object is return due the nature of the function.
}
\description{
You may supply manually your username, password and x-api-key, or store them as values at the renviron document. For this, use the function \code{usethis::edit_r_environ()} and add the following keys: \code{plugfield_username}, \code{plugfield_password} and \code{plugfield_x_api_key}.
}
